updated the files readme and platform_data
updated the files
readme.txt
platform_data.py
and deleted the earlier ones present in the folder
